requirements.txt
asgiref==3.2.5
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==3.0.4
django-grappelli==2.14.1
django-nested-admin==3.2.4
django-phonenumber-field==4.0.0
environs==7.3.1
idna==2.9
marshmallow==3.5.1
oauthlib==3.1.0
phonenumbers==8.12.1
Pillow==7.1.0
pycparser==2.20
python-dotenv==0.12.0
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.2
sqlparse==0.3.1
urllib3==1.25.8
